laravel 6.0 bigInteger support

+++ src/Dmocho/Updater/UpdaterBlueprint.php
@@ -1,4 +1,4 @@
-<?php 
+<?php
 
 namespace Dmocho\Updater;
 
@@ -6,8 +6,8 @@ class UpdaterBlueprint extends \Illuminate\Database\Schema\Blueprint
 {
     public function authors()
     {
-        $this->integer(config('updater.blueprint.created_by'))->unsigned()->nullable();
-        $this->integer(config('updater.blueprint.updated_by'))->unsigned()->nullable();
+        $this->bigInteger(config('updater.blueprint.created_by'))->unsigned()->nullable();
+        $this->bigInteger(config('updater.blueprint.updated_by'))->unsigned()->nullable();
 
         $this->foreign(config('updater.blueprint.created_by'))->references('id')->on('users')->onUpdate('cascade')->onDelete('set null');
         $this->foreign(config('updater.blueprint.updated_by'))->references('id')->on('users')->onUpdate('cascade')->onDelete('set null');
@@ -15,7 +15,7 @@ class UpdaterBlueprint extends \Illuminate\Database\Schema\Blueprint
 
     public function softDeleters()
     {
-        $this->integer(config('updater.blueprint.deleted_by'))->unsigned()->nullable();
+        $this->bigInteger(config('updater.blueprint.deleted_by'))->unsigned()->nullable();
         $this->foreign(config('updater.blueprint.deleted_by'))->references('id')->on('users')->onUpdate('cascade')->onDelete('set null');
     }
